switch from float to double

+++ Analog4Poles.h
@@ -4,24 +4,24 @@
 class Analog4Poles {
 
 private:
-	float pole1;
-	float pole2;
-	float pole3;
-	float pole4;
-	float feed1;
-	float feed2;
+	double pole1;
+	double pole2;
+	double pole3;
+	double pole4;
+	double feed1;
+	double feed2;
 
-	float resLevel;
-	float modMax;
-	float modDepth;
-	float modScale;
-	float levCorrect;
+	double resLevel;
+	double modMax;
+	double modDepth;
+	double modScale;
+	double levCorrect;
 
 public:
-	float reso;
-	float dist;
-	float freq;
-	float out;
+	double reso;
+	double dist;
+	double freq;
+	double out;
 
 	Analog4Poles() {
 		pole1 =  0.0;
@@ -43,9 +43,9 @@ public:
 		out   =  0.0;
 	}
 
-	float tic(float in) {
+	double tic(double in) {
 
-		float s = freq; // TODO: logarithmic curve
+		double s = freq; // TODO: logarithmic curve
 
 		pole4 = pole3 * s + pole4 * (1 - s); // pole4 = pole4Filter(pole3);
 		pole3 = pole2 * s + pole3 * (1 - s); // pole3 = pole3Filter(pole2);
