+++ constants.h
@@ -5,6 +5,6 @@
 #define SCREEN_W 720
 #define SCREEN_H 600
 
-const float SAMPLE_RATE = 44100.0;
+const double SAMPLE_RATE = 44100.0;
 
 #endif
